break it up into pieces
